move AbsoluteMouseAPI
